+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/guice/ProxyModule.java
@@ -21,7 +21,6 @@ import stroom.proxy.app.Config;
 import stroom.proxy.app.ContentSyncService;
 import stroom.proxy.app.ProxyConfigHealthCheck;
 import stroom.proxy.app.ProxyConfigHolder;
-import stroom.proxy.app.ProxyPathConfig;
 import stroom.proxy.app.RestClientConfig;
 import stroom.proxy.app.handler.ForwardStreamHandlerFactory;
 import stroom.proxy.app.handler.ProxyRequestHandler;
@@ -48,7 +47,6 @@ import stroom.security.api.SecurityContext;
 import stroom.security.mock.MockSecurityContext;
 import stroom.task.impl.TaskContextModule;
 import stroom.util.BuildInfoProvider;
-import stroom.util.config.ConfigLocation;
 import stroom.util.entityevent.EntityEventBus;
 import stroom.util.guice.FilterBinder;
 import stroom.util.guice.FilterInfo;
@@ -59,7 +57,6 @@ import stroom.util.guice.ServletBinder;
 import stroom.util.io.BufferFactory;
 import stroom.util.io.HomeDirProvider;
 import stroom.util.io.HomeDirProviderImpl;
-import stroom.util.io.PathConfig;
 import stroom.util.io.PathCreator;
 import stroom.util.io.TempDirProvider;
 import stroom.util.io.TempDirProviderImpl;
@@ -70,12 +67,14 @@ import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import io.dropwizard.client.JerseyClientBuilder;
 import io.dropwizard.client.JerseyClientConfiguration;
+import io.dropwizard.client.ssl.TlsConfiguration;
 import io.dropwizard.lifecycle.Managed;
 import io.dropwizard.setup.Environment;
 import org.glassfish.jersey.logging.LoggingFeature;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.nio.file.Path;
 import java.util.Optional;
 import javax.inject.Provider;
@@ -185,7 +184,8 @@ public class ProxyModule extends AbstractModule {
     @Singleton
     Client provideJerseyClient(final RestClientConfig restClientConfig,
                                final Environment environment,
-                               final Provider<BuildInfo> buildInfoProvider) {
+                               final Provider<BuildInfo> buildInfoProvider,
+                               final PathCreator pathCreator) {
 
         // RestClientConfig is really just JerseyClientConfiguration
         final JerseyClientConfiguration jerseyClientConfiguration = restClientConfig;
@@ -195,11 +195,44 @@ public class ProxyModule extends AbstractModule {
         if (jerseyClientConfiguration.getUserAgent().isEmpty()) {
             final String userAgent = PROXY_JERSEY_CLIENT_USER_AGENT_PREFIX
                     + buildInfoProvider.get().getBuildVersion();
-            LOGGER.info("Setting jersey client user agent string to [{}]", userAgent);
+            LOGGER.info("Setting rest client user agent string to [{}]", userAgent);
             jerseyClientConfiguration.setUserAgent(Optional.of(userAgent));
         }
 
-        LOGGER.info("Creating jersey client {}", PROXY_JERSEY_CLIENT_NAME);
+        // Mutating the TLS config is not ideal but I'm not sure there is another way.
+        // We need to allow for relative paths (relative to proxy home), '~', and other system
+        // props in the path. Therefore if path creator produces a different path to what was
+        // configured then update the config object.
+        final TlsConfiguration tlsConfiguration = jerseyClientConfiguration.getTlsConfiguration();
+        if (tlsConfiguration != null) {
+            if (tlsConfiguration.getKeyStorePath() != null) {
+                final File modifiedKeyStorePath = new File(pathCreator.makeAbsolute(
+                        pathCreator.replaceSystemProperties(
+                                tlsConfiguration.getKeyStorePath().getPath())));
+
+                if (!modifiedKeyStorePath.getPath().equals(tlsConfiguration.getKeyStorePath().getPath())) {
+                    LOGGER.info("Updating rest client key store path from {} to {}",
+                            tlsConfiguration.getKeyStorePath(),
+                            modifiedKeyStorePath);
+                    tlsConfiguration.setKeyStorePath(modifiedKeyStorePath);
+                }
+            }
+
+            if (tlsConfiguration.getTrustStorePath() != null) {
+                final File modifiedTrustStorePath = new File(pathCreator.makeAbsolute(
+                        pathCreator.replaceSystemProperties(
+                                tlsConfiguration.getTrustStorePath().getPath())));
+
+                if (!modifiedTrustStorePath.getPath().equals(tlsConfiguration.getTrustStorePath().getPath())) {
+                    LOGGER.info("Updating rest client trust store path from {} to {}",
+                            tlsConfiguration.getTrustStorePath(),
+                            modifiedTrustStorePath);
+                    tlsConfiguration.setTrustStorePath(modifiedTrustStorePath);
+                }
+            }
+        }
+
+        LOGGER.info("Creating jersey rest client {}", PROXY_JERSEY_CLIENT_NAME);
         return new JerseyClientBuilder(environment)
                 .using(jerseyClientConfiguration)
                 .build(PROXY_JERSEY_CLIENT_NAME)