+++ stroom-util/src/main/java/stroom/util/validation/ValidFilePathValidatorImpl.java
@@ -1,13 +1,27 @@
 package stroom.util.validation;
 
+import stroom.util.io.PathCreator;
 import stroom.util.shared.validation.ValidFilePath;
 import stroom.util.shared.validation.ValidFilePathValidator;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.nio.file.Files;
 import java.nio.file.Path;
+import javax.inject.Inject;
 import javax.validation.ConstraintValidatorContext;
 
 public class ValidFilePathValidatorImpl implements ValidFilePathValidator {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ValidFilePathValidatorImpl.class);
+
+    private final PathCreator pathCreator;
+
+    @Inject
+    public ValidFilePathValidatorImpl(final PathCreator pathCreator) {
+        this.pathCreator = pathCreator;
+    }
 
     /**
      * Initializes the validator in preparation for
@@ -32,16 +46,32 @@ public class ValidFilePathValidatorImpl implements ValidFilePathValidator {
      * This method can be accessed concurrently, thread-safety must be ensured
      * by the implementation.
      *
-     * @param value   object to validate
+     * @param file    object to validate
      * @param context context in which the constraint is evaluated
      * @return {@code false} if {@code value} does not pass the constraint
      */
     @Override
-    public boolean isValid(final String value, final ConstraintValidatorContext context) {
-        if (value != null) {
-            return Files.isRegularFile(Path.of(value));
+    public boolean isValid(final String file, final ConstraintValidatorContext context) {
+        final boolean isValid;
+        if (file != null) {
+            // Use the PathCreator so we can interpret relative paths and paths with '~' in.
+            final String modifiedFile = pathCreator.makeAbsolute(pathCreator.replaceSystemProperties(file));
+
+            LOGGER.debug("Validating file {} (modified to {})", file, modifiedFile);
+            isValid = Files.isRegularFile(Path.of(modifiedFile));
+            if (!isValid) {
+                String msg = context.getDefaultConstraintMessageTemplate();
+                if (!modifiedFile.equals(file)) {
+                    msg += " (as absolute path: [" + modifiedFile + "]";
+                }
+                context.disableDefaultConstraintViolation();
+                context
+                        .buildConstraintViolationWithTemplate(msg)
+                        .addConstraintViolation();
+            }
         } else {
-            return true;
+            isValid = true;
         }
+        return isValid;
     }
 }